Move remaining namespace tests from runtests.c to ns_tests.c
