+++ expat/tests/handlers.h
@@ -320,6 +320,12 @@ extern int XMLCALL external_entity_faulter2(XML_Parser parser,
                                             const XML_Char *systemId,
                                             const XML_Char *publicId);
 
+extern int XMLCALL external_entity_handler(XML_Parser parser,
+                                           const XML_Char *context,
+                                           const XML_Char *base,
+                                           const XML_Char *systemId,
+                                           const XML_Char *publicId);
+
 /* Declaration handlers for entity declarations */
 extern void XMLCALL entity_suspending_decl_handler(void *userData,
                                                    const XML_Char *name,
